TIP-1050: move classes
