+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/cursors.yml
@@ -61,14 +61,14 @@ services:
 
     akeneo.pim.enrichment.factory.product_identifier_cursor_from_size_cursor:
         public: false
-        class: 'Akeneo\Pim\Enrichment\Bundle\Cursor\FromSizeIdentifierResultCursorFactory'
+        class: 'Akeneo\Pim\Enrichment\Bundle\Elasticsearch\FromSizeIdentifierResultCursorFactory'
         arguments:
         - '@akeneo_elasticsearch.client.product'
         - 'pim_catalog_product'
 
     akeneo.pim.enrichment.factory.product_identifier_cursor_search_after_size_cursor:
         public: false
-        class: 'Akeneo\Pim\Enrichment\Bundle\Storage\ElasticsearchAndSql\ProductGrid\SearchAfterSizeIdentifierResultCursorFactory'
+        class: 'Akeneo\Pim\Enrichment\Bundle\Elasticsearch\SearchAfterSizeIdentifierResultCursorFactory'
         arguments:
             - '@akeneo_elasticsearch.client.product'
             - 'pim_catalog_product'
